+++ randovania/gui/game_details/game_details_window.py
@@ -64,6 +64,13 @@ class GameDetailsWindow(CloseEventWidget, Ui_GameDetailsWindow, BackgroundTaskMi
         self._window_manager = window_manager
         self._game_details_tabs = []
 
+        self.progress_bar.setVisible(False)
+        self.stop_background_process_button.setVisible(False)
+
+        self.status_bar.addWidget(self.progress_label)
+        self.status_bar.addPermanentWidget(self.progress_bar)
+        self.status_bar.addPermanentWidget(self.stop_background_process_button)
+
         # Ui
         self._tool_button_menu = QtWidgets.QMenu(self.tool_button)
         self.tool_button.setMenu(self._tool_button_menu)
@@ -317,11 +324,18 @@ class GameDetailsWindow(CloseEventWidget, Ui_GameDetailsWindow, BackgroundTaskMi
                 )
 
     def enable_buttons_with_background_tasks(self, value: bool):
+        self.stop_background_process_button.setVisible(True)
         self.stop_background_process_button.setEnabled(not value and self._can_stop_background_process)
         self.export_iso_button.setEnabled(value)
         generator_frontend.export_busy = not value
 
+        if not self._can_stop_background_process:
+            self.stop_background_process_button.setToolTip("This game doesn't let you stop the export.")
+        else:
+            self.stop_background_process_button.setToolTip("")
+
     def update_progress(self, message: str, percentage: int):
+        self.progress_bar.setVisible(True)
         self.progress_label.setText(message)
         if "Aborted" in message:
             percentage = 0